requirements.txt
Faker==6.1.1
apache-beam==2.27.0
google-cloud-pubsub==2.3.0
rich==9.10.0
